split out the charts into components
